ipv6: drop 1.12+ variant of @ipv6_add_static_address_manually_not_active
NM 1.12 still adds the IPv6LL address when an IPv6 address is added
externally. The reason why this test passed is that it triggered an
assertion:

 src/devices/nm-device.c: line 8080
 (nm_device_get_configured_mtu_for_wired): should not be reached

which caused a restart of the NM process and therefore the LL address
wasn't added.

Once the assertion is fixed, the LL gets added because the following
happens:

 - NM detects that the address is added externally

 - in queued_ip_config_change() we call check_and_add_ipv6ll_addr()
   for the reason explained by this code comment:

  "If no IPv6 link-local address exists but other addresses do then we
   must add the LL address to remain conformant with RFC 3513 chapter 2.1
   ("Addressing Model"): "All interfaces are required to have at least
   one link-local unicast address".

 - check_and_add_ipv6ll_addr() generates and adds a LL address

 - then, an in-memory connection is created and device becomes
   "external". From this point NM no longer touches the interface.

I think changing this behavior and requiring the user to add a LL
address (or change the addrgenmode and cycle the interface down+up)
would be very user-unfriendly and so I prefer we keep adding address
even if the connection is going to become "external".