add 'policy_based_routing_doc_procedure' test
Add test, that covers use case from documentation example.

It creates 4 namespaces (4 different networks), two are providers, one 
is for internal workstations and one for servers. The test cretaes 
policy-based routing rules in NM connections, and then checks, that 
everything works (servers use privder A, workstations provider B).

To make this work, eth0 has to be disabled, and so there is no access to 
internet. The fix is to create veths in providers' namespaces and assign 
them the same address (172.20.20.20). Then check tcpdump log, whether 
there is the traffic in correct namespace.