version_control: support multiple version tags
Imagine a bugfix or feature gets backported to older versions.
For example, it is now available in @ver+=1.26.6 @ver+=1.28.4 @ver+=1.29.4
That means, it works not on 1.28.2, but it would work on 1.28.6.

Extend the version comparison to account for multiple version tags, in a
way that makes sense and is suitable to express such conditions.

Note that there is also a change in behavior: ver+1.28 would previously
also accept "1.28.2". Now, this is the same as ver+=1.29. Howver, no
"ver+1.*" tags are currently used, so it doesn't matter.